requirements.txt
pandas
numpy
scikit-learn
matplotlib
loguru
mkdocs
pip
pytest
python-dotenv
ruff
tqdm
typer
fastapi
pillow
uvicorn
pydantic
pydantic_settings
-e .
